Code Cleanup * About page now a scroll view * Updated Rover assets - rounded icons
