+++ NAPIGallery/Assets.xcassets/Rover_Spirit_Icon.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "NASA_Mars_Rover_Spirit_Small@1x.jpg",
+      "filename" : "Mars_Rover_Spirit_Small_Rounded@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "NASA_Mars_Rover_Spirit_Small@2x.jpg",
+      "filename" : "Mars_Rover_Spirit_Small_Rounded@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "NASA_Mars_Rover_Spirit_Small@3x.jpg",
+      "filename" : "Mars_Rover_Spirit_Small_Rounded@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
