+++ NAPIGallery/Assets.xcassets/dark_rift.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "Mars_Rover_Opportunity_Small_Rounded@1x.png",
+      "filename" : "darkrift-orig_full@1x.jpg",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "Mars_Rover_Opportunity_Small_Rounded@2x.png",
+      "filename" : "darkrift-orig_full@2x.jpg",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "Mars_Rover_Opportunity_Small_Rounded@3x.png",
+      "filename" : "darkrift-orig_full@3x.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
