+++ NAPIGallery/System/NAPIAboutContentView.swift
@@ -12,17 +12,22 @@ struct NAPIAboutContentView: View {
     
     var body: some View {
         GeometryReader { metrics in
+            ZStack {
                 Image("About_Background")
                     .resizable()
                     .aspectRatio(contentMode: .fill)
-                    .edgesIgnoringSafeArea(/*@START_MENU_TOKEN@*/.all/*@END_MENU_TOKEN@*/)
+                    .edgesIgnoringSafeArea(.all)
                     .frame(width: metrics.size.width,
                            height: metrics.size.height,
                            alignment: /*@START_MENU_TOKEN@*/.center/*@END_MENU_TOKEN@*/)
-                    .overlay(info)
             }
+            ScrollView {
+                info
+            }
+            .frame(width: metrics.size.width, height: metrics.size.height, alignment: .center)
+        }
     }
-    
+
     var info: some View {
         VStack {
             Text("Welcome to the\nNASA API Gallery")
@@ -30,13 +35,14 @@ struct NAPIAboutContentView: View {
                 .minimumScaleFactor(0.5)
                 .multilineTextAlignment(.center)
                 .foregroundColor(.white)
-                .padding(.top, 50)
-            Spacer()
+                .padding(.top)
+            Spacer(minLength: 25.0)
             Text("This app will allow you to view the various images and other data. The information is publicly available from the NASA API but this app makes it easy to retrieve and visualize the information!")
                 .font(.title2)
                 .multilineTextAlignment(.leading)
                 .padding()
                 .foregroundColor(.white)
+            Spacer(minLength: 25.0)
             infoWithLink
                 .font(.title2)
                 .multilineTextAlignment(.leading)
@@ -45,13 +51,12 @@ struct NAPIAboutContentView: View {
                 .onTapGesture {
                     openURL(URL(string: "https://api.nasa.gov")!)
                 }
-            Spacer()
+            Spacer(minLength: 50.0)
             Text("You can use the DEMO_KEY provided but it will only allow for a limited number of data requests (30 requests/hr, 50 requests/day.")
                 .multilineTextAlignment(.leading)
                 .font(Font.caption.weight(.bold))
                 .padding()
                 .foregroundColor(.white)
-            Spacer()
         }
     }
     
@@ -60,7 +65,7 @@ struct NAPIAboutContentView: View {
                 + Text("https://api.nasa.gov ")
                     .foregroundColor(.yellow)
                     .underline(color: .white))
-                + Text("to generate a free API Key which will allow for more data requests per hour"))
+                + Text("to generate a free API Key which will allow for more data requests per hour."))
     }
 }
